+++ ios/threadable_better_player/Sources/ThreadableBetterPlayer/BetterPlayerView.swift
@@ -1,0 +1,17 @@
+import AVFoundation
+import UIKit
+
+final class BetterPlayerView: UIView {
+    override class var layerClass: AnyClass {
+        AVPlayerLayer.self
+    }
+
+    var playerLayer: AVPlayerLayer {
+        layer as! AVPlayerLayer
+    }
+
+    var player: AVPlayer? {
+        get { playerLayer.player }
+        set { playerLayer.player = newValue }
+    }
+}